Expose the defanging logic as a library so it can be embedded and compiled to wasm (#13)
